// frontend/src/api/generated/models/alertHistoryResponseSeverity.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */

export type AlertHistoryResponseSeverity = typeof AlertHistoryResponseSeverity[keyof typeof AlertHistoryResponseSeverity];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AlertHistoryResponseSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;