// frontend/src/api/generated/models/alertHistoryResponseStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */

export type AlertHistoryResponseStatus = typeof AlertHistoryResponseStatus[keyof typeof AlertHistoryResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AlertHistoryResponseStatus = {
  pending: 'pending',
  sent: 'sent',
  acknowledged: 'acknowledged',
  resolved: 'resolved',
} as const;